requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
JPype1==1.5.0
kombu==5.3.7
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pillow==10.3.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
six==1.16.0
sqlparse==0.4.4
tabula-py==2.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
